@@ -1,26 +1,27 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Nullable>enable</Nullable>
+        <LangVersion>10</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-        <Compile Remove="Serialization\**"/>
-        <EmbeddedResource Remove="Serialization\**"/>
-        <None Remove="Serialization\**"/>
+        <Compile Remove="Serialization\**" />
+        <EmbeddedResource Remove="Serialization\**" />
+        <None Remove="Serialization\**" />
     </ItemGroup>
 
     <ItemGroup>
-        <Compile Remove="SerializerFactory.cs"/>
+        <Compile Remove="SerializerFactory.cs" />
     </ItemGroup>
 
     <ItemGroup>
-        <None Include="SerializerFactory.cs"/>
+        <None Include="SerializerFactory.cs" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Tauron.Application.Common\Tauron.Application.Common.csproj"/>
+        <ProjectReference Include="..\Tauron.Application.Common\Tauron.Application.Common.csproj" />
     </ItemGroup>
 
     <ItemGroup>
@@ -36,8 +37,8 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="SmartAnalyzers.ExceptionAnalyzer" Version="1.0.10"/>
-        <PackageReference Include="SmartAnalyzers.MultithreadingAnalyzer" Version="1.1.31"/>
+        <PackageReference Include="SmartAnalyzers.ExceptionAnalyzer" Version="1.0.10" />
+        <PackageReference Include="SmartAnalyzers.MultithreadingAnalyzer" Version="1.1.31" />
     </ItemGroup>
 
 </Project>
