@@ -1,7 +1,6 @@
 ﻿using JetBrains.Annotations;
 
-namespace Tauron.Application.Files.HeaderedText
-{
-    [PublicAPI]
-    public sealed record ContextEntry(string Key, string Content);
-}
+namespace Tauron.Application.Files.HeaderedText;
+
+[PublicAPI]
+public sealed record ContextEntry(string Key, string Content);
