@@ -1,27 +1,27 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using JetBrains.Annotations;
 using Tauron.Application.VirtualFiles;
 
-namespace Tauron.Application.Files
+namespace Tauron.Application.Files;
+
+[PublicAPI]
+public static class IoExtensions
 {
-    [PublicAPI]
-    public static class IoExtensions
+    public static void WriteAllBytes(this IFile file, byte[] bytes)
     {
-        public static void WriteAllBytes(this IFile file, byte[] bytes)
-        {
-            using var write = file.Create();
-            write.Write(bytes, 0, bytes.Length);
-        }
+        using var write = file.Open(FileAccess.Write);
+        write.Write(bytes, 0, bytes.Length);
+    }
 
-        public static byte[] ReadAllBytes(this IFile file)
-        {
-            if (!file.Exist) return new byte[0];
+    public static byte[] ReadAllBytes(this IFile file)
+    {
+        if (!file.Exist) return Array.Empty<byte>();
 
-            using var stream = file.Open(FileAccess.Read);
-            byte[] arr = new byte[stream.Length];
-            stream.Read(arr, 0, arr.Length);
+        using var stream = file.Open(FileAccess.Read);
+        var arr = new byte[stream.Length];
+        stream.Read(arr, 0, arr.Length);
 
-            return arr;
-        }
+        return arr;
     }
 }
