@@ -1,7 +1,6 @@
 ﻿using System;
 
-namespace Tauron.Application.Files.Ini
-{
-    [Serializable]
-    public abstract record IniEntry(string Key);
-}
+namespace Tauron.Application.Files.Ini;
+
+[Serializable]
+public abstract record IniEntry(string Key);
