@@ -2,45 +2,44 @@
 using System.IO;
 using JetBrains.Annotations;
 
-namespace Tauron.Application.Files.Ini.Parser
+namespace Tauron.Application.Files.Ini.Parser;
+
+[PublicAPI]
+public class IniWriter
 {
-    [PublicAPI]
-    public class IniWriter
+    private readonly IniFile _file;
+    private readonly TextWriter _writer;
+
+    public IniWriter(IniFile file, TextWriter writer)
     {
-        private readonly IniFile _file;
-        private readonly TextWriter _writer;
+        _file = file;
+        _writer = writer;
+    }
 
-        public IniWriter(IniFile file, TextWriter writer)
+    public void Write()
+    {
+        try
         {
-            _file = file;
-            _writer = writer;
+            foreach (var section in _file)
+            {
+                _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "[{0}]", section.Name));
+                foreach (var iniEntry in section)
+                    if (iniEntry is SingleIniEntry(var key, var data))
+                    {
+                        _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "{0}={1}", key, data));
+                    }
+                    else
+                    {
+                        var (name, values) = (ListIniEntry)iniEntry;
+                        foreach (var value in values)
+                            _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "{0}={1}", name, value));
+                    }
+            }
         }
-
-        public void Write()
+        finally
         {
-            try
-            {
-                foreach (var section in _file)
-                {
-                    _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "[{0}]", section.Name));
-                    foreach (var iniEntry in section)
-                        if (iniEntry is SingleIniEntry(var key, var data))
-                        {
-                            _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "{0}={1}", key, data));
-                        }
-                        else
-                        {
-                            var (name, values) = (ListIniEntry)iniEntry;
-                            foreach (var value in values)
-                                _writer.WriteLine(string.Format(CultureInfo.InvariantCulture, "{0}={1}", name, value));
-                        }
-                }
-            }
-            finally
-            {
-                _writer.Flush();
-                _writer.Dispose();
-            }
+            _writer.Flush();
+            _writer.Dispose();
         }
     }
 }
