@@ -1,9 +1,8 @@
 ﻿using System;
 using JetBrains.Annotations;
 
-namespace Tauron.Application.Files.Ini
-{
-    [PublicAPI]
-    [Serializable]
-    public sealed record SingleIniEntry(string Key, string? Value) : IniEntry(Key);
-}
+namespace Tauron.Application.Files.Ini;
+
+[PublicAPI]
+[Serializable]
+public sealed record SingleIniEntry(string Key, string? Value) : IniEntry(Key);
