@@ -1,171 +1,169 @@
 using System;
 using System.IO;
 using JetBrains.Annotations;
+using Tauron.Application.VirtualFiles;
 
-namespace Tauron.Application.Files.Json
+namespace Tauron.Application.Files.Json;
+
+[PublicAPI]
+public abstract partial class JsonNode
 {
-    [PublicAPI]
-    public abstract partial class JsonNode
+    public abstract void SerializeBinary(BinaryWriter aWriter);
+
+    public void SaveToBinaryStream(Stream aData)
     {
-        public abstract void SerializeBinary(BinaryWriter aWriter);
+        SerializeBinary(new BinaryWriter(aData));
+    }
 
-        public void SaveToBinaryStream(Stream aData)
+
+    public void SaveToBinaryFile(IFile aFile)
+    {
+        using var fileStream = aFile.Open(FileAccess.Write);
+        SaveToBinaryStream(fileStream);
+    }
+
+    public string SaveToBinaryBase64()
+    {
+        using var stream = new MemoryStream();
+        SaveToBinaryStream(stream);
+        stream.Position = 0;
+
+        return Convert.ToBase64String(stream.ToArray());
+    }
+
+    public static JsonNode DeserializeBinary(BinaryReader aReader)
+    {
+        var type = (JsonNodeType)aReader.ReadByte();
+        switch (type)
         {
-            SerializeBinary(new BinaryWriter(aData));
-        }
-
-
-        public void SaveToBinaryFile(string aFileName)
-        {
-            aFileName.CreateDirectoryIfNotExis();
-
-            using var fileStream = File.OpenWrite(aFileName);
-            SaveToBinaryStream(fileStream);
-        }
-
-        public string SaveToBinaryBase64()
-        {
-            using var stream = new MemoryStream();
-            SaveToBinaryStream(stream);
-            stream.Position = 0;
-
-            return Convert.ToBase64String(stream.ToArray());
-        }
-
-        public static JsonNode DeserializeBinary(BinaryReader aReader)
-        {
-            var type = (JsonNodeType)aReader.ReadByte();
-            switch (type)
+            case JsonNodeType.Array:
             {
-                case JsonNodeType.Array:
-                {
-                    var count = aReader.ReadInt32();
-                    var tmp = new JsonArray();
-                    for (var i = 0; i < count; i++) tmp.Add(DeserializeBinary(aReader));
+                var count = aReader.ReadInt32();
+                var tmp = new JsonArray();
+                for (var i = 0; i < count; i++) tmp.Add(DeserializeBinary(aReader));
 
-                    return tmp;
-                }
-                case JsonNodeType.Object:
-                {
-                    var count = aReader.ReadInt32();
-                    var tmp = new JsonObject();
-                    for (var i = 0; i < count; i++)
-                    {
-                        var key = aReader.ReadString();
-                        var val = DeserializeBinary(aReader);
-                        tmp.Add(key, val);
-                    }
-
-                    return tmp;
-                }
-                case JsonNodeType.String:
-                {
-                    return new JsonString(aReader.ReadString());
-                }
-                case JsonNodeType.Number:
-                {
-                    return new JsonNumber(aReader.ReadDouble());
-                }
-                case JsonNodeType.Boolean:
-                {
-                    return new JsonBool(aReader.ReadBoolean());
-                }
-                case JsonNodeType.NullValue:
-                {
-                    return JsonNull.CreateOrGet();
-                }
-                default:
-                {
-                    throw new InvalidOperationException("Error deserializing JSON. Unknown tag: " + type);
-                }
+                return tmp;
             }
-        }
-
-        public static JsonNode LoadFromBinaryStream(Stream aData)
-        {
-            using var binaryReader = new BinaryReader(aData);
-
-            return DeserializeBinary(binaryReader);
-        }
-
-        public static JsonNode LoadFromBinaryFile(string aFileName)
-        {
-            using var fileStream = File.OpenRead(aFileName);
-
-            return LoadFromBinaryStream(fileStream);
-        }
-
-        public static JsonNode LoadFromBinaryBase64(string aBase64)
-        {
-            var tmp = Convert.FromBase64String(aBase64);
-            var stream = new MemoryStream(tmp) { Position = 0 };
-
-            return LoadFromBinaryStream(stream);
-        }
-    }
-
-    public partial class JsonArray
-    {
-        public override void SerializeBinary(BinaryWriter aWriter)
-        {
-            aWriter.Write((byte)JsonNodeType.Array);
-            aWriter.Write(_list.Count);
-            foreach (var jsonNode in _list)
-                jsonNode.SerializeBinary(aWriter);
-        }
-    }
-
-    public partial class JsonObject
-    {
-        public override void SerializeBinary(BinaryWriter aWriter)
-        {
-            aWriter.Write((byte)JsonNodeType.Object);
-            aWriter.Write(_dict.Count);
-            foreach (var key in _dict.Keys)
+            case JsonNodeType.Object:
             {
-                aWriter.Write(key);
-                _dict[key].SerializeBinary(aWriter);
+                var count = aReader.ReadInt32();
+                var tmp = new JsonObject();
+                for (var i = 0; i < count; i++)
+                {
+                    var key = aReader.ReadString();
+                    var val = DeserializeBinary(aReader);
+                    tmp.Add(key, val);
+                }
+
+                return tmp;
+            }
+            case JsonNodeType.String:
+            {
+                return new JsonString(aReader.ReadString());
+            }
+            case JsonNodeType.Number:
+            {
+                return new JsonNumber(aReader.ReadDouble());
+            }
+            case JsonNodeType.Boolean:
+            {
+                return new JsonBool(aReader.ReadBoolean());
+            }
+            case JsonNodeType.NullValue:
+            {
+                return JsonNull.CreateOrGet();
+            }
+            default:
+            {
+                throw new InvalidOperationException("Error deserializing JSON. Unknown tag: " + type);
             }
         }
     }
 
-    public partial class JsonString
+    public static JsonNode LoadFromBinaryStream(Stream aData)
     {
-        public override void SerializeBinary(BinaryWriter aWriter)
-        {
-            aWriter.Write((byte)JsonNodeType.String);
-            aWriter.Write(_data);
-        }
+        using var binaryReader = new BinaryReader(aData);
+
+        return DeserializeBinary(binaryReader);
     }
 
-    public partial class JsonNumber
+    public static JsonNode LoadFromBinaryFile(string aFileName)
     {
-        public override void SerializeBinary(BinaryWriter aWriter)
-        {
-            aWriter.Write((byte)JsonNodeType.Number);
-            aWriter.Write(_data);
-        }
+        using var fileStream = File.OpenRead(aFileName);
+
+        return LoadFromBinaryStream(fileStream);
     }
 
-    public partial class JsonBool
+    public static JsonNode LoadFromBinaryBase64(string aBase64)
     {
-        public override void SerializeBinary(BinaryWriter aWriter)
+        var tmp = Convert.FromBase64String(aBase64);
+        var stream = new MemoryStream(tmp) { Position = 0 };
+
+        return LoadFromBinaryStream(stream);
+    }
+}
+
+public partial class JsonArray
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
+    {
+        aWriter.Write((byte)JsonNodeType.Array);
+        aWriter.Write(_list.Count);
+        foreach (var jsonNode in _list)
+            jsonNode.SerializeBinary(aWriter);
+    }
+}
+
+public partial class JsonObject
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
+    {
+        aWriter.Write((byte)JsonNodeType.Object);
+        aWriter.Write(_dict.Count);
+        foreach (var key in _dict.Keys)
         {
-            aWriter.Write((byte)JsonNodeType.Boolean);
-            aWriter.Write(_data);
+            aWriter.Write(key);
+            _dict[key].SerializeBinary(aWriter);
         }
     }
+}
 
-    public partial class JsonNull
+public partial class JsonString
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
     {
-        public override void SerializeBinary(BinaryWriter aWriter)
-        {
-            aWriter.Write((byte)JsonNodeType.NullValue);
-        }
+        aWriter.Write((byte)JsonNodeType.String);
+        aWriter.Write(_data);
     }
+}
 
-    internal partial class JsonLazyCreator
+public partial class JsonNumber
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
     {
-        public override void SerializeBinary(BinaryWriter aWriter) { }
+        aWriter.Write((byte)JsonNodeType.Number);
+        aWriter.Write(_data);
     }
+}
+
+public partial class JsonBool
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
+    {
+        aWriter.Write((byte)JsonNodeType.Boolean);
+        aWriter.Write(_data);
+    }
+}
+
+public partial class JsonNull
+{
+    public override void SerializeBinary(BinaryWriter aWriter)
+    {
+        aWriter.Write((byte)JsonNodeType.NullValue);
+    }
+}
+
+internal partial class JsonLazyCreator
+{
+    public override void SerializeBinary(BinaryWriter aWriter) { }
 }
