@@ -14,8 +14,8 @@ public sealed record PathInfo(string Path, PathType Kind)
 {
 
     public static implicit operator string(PathInfo path)
-        => path.Path;
+        => GenericPathHelper.ToRelativePath(path).Path;
 
     public static implicit operator PathInfo(string path) 
-        => new(GenericPathHelper.NormalizePath(path), GenericPathHelper.IsAbsolute(path) ? PathType.Absolute : PathType.Relative);
+        => new(GenericPathHelper.NormalizeString(path), GenericPathHelper.IsAbsolute(path) ? PathType.Absolute : PathType.Relative);
 }
