@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.IO.Compression;
-using System.Linq;
 using System.Text;
 using JetBrains.Annotations;
+using Tauron.Application.VirtualFiles;
 
 namespace Tauron;
 
@@ -14,7 +12,86 @@ namespace Tauron;
 [PublicAPI]
 public static class IoExtensions
 {
-    public static void AddFilesFromDictionary(this ZipArchive destination, string sourceDirectoryName)
+    public static string PathShorten(this string path, int length)
+    {
+        var pathParts = path.Split('\\');
+        var pathBuild = new StringBuilder(path.Length);
+        var lastPart = pathParts[^1];
+        var prevPath = "";
+
+        //Erst prüfen ob der komplette String evtl. bereits kürzer als die Maximallänge ist
+        if (path.Length >= length) return path;
+
+        for (var i = 0; i < pathParts.Length - 1; i++)
+        {
+            pathBuild.Append(pathParts[i] + @"\");
+
+            if ((pathBuild + @"...\" + lastPart).Length >= length) return prevPath;
+
+            prevPath = pathBuild + @"...\" + lastPart;
+        }
+
+        return prevPath;
+    }
+    
+    public static IEnumerable<string> EnumerateTextLinesIfExis(this IFile file)
+    {
+        if (!file.Exist) yield break;
+
+        using var reader = new StreamReader(file.Open(FileAccess.Read));
+
+        while (true)
+        {
+            var line = reader.ReadLine();
+
+            if (line is null) yield break;
+
+            yield return line;
+        }
+    }
+    
+    public static IEnumerable<string> EnumerateTextLines(this TextReader reader)
+    {
+        while (true)
+        {
+            var line = reader.ReadLine();
+
+            if (line == null) yield break;
+
+            yield return line;
+        }
+    }
+    
+    public static void Clear(this IDirectory dic)
+    {
+        if (dic is null) throw new ArgumentNullException(nameof(dic));
+
+        if (!dic.Exist) return;
+
+        if(dic is IHasFileAttributes dicAttr)
+            dicAttr.Attributes = FileAttributes.Normal;
+
+        foreach (var file in dic.Files)
+        {
+            if (file is IHasFileAttributes fileAttributes)
+                fileAttributes.Attributes = FileAttributes.Normal;
+            file.Delete();
+        }
+
+        foreach (var directory in dic.Directories)
+        {
+            Clear(directory);
+            directory.Delete();
+        }
+    }
+
+    public static TextWriter OpenWrite(this IFile file)
+        => new StreamWriter(file.Open(FileAccess.Write));
+
+    public static TextReader OpenRead(this IFile file)
+        => new StreamReader(file.Open(FileAccess.Read));
+    
+    /*public static void AddFilesFromDictionary(this ZipArchive destination, string sourceDirectoryName)
     {
         var stack = new Stack<FileSystemInfo>();
         new DirectoryInfo(sourceDirectoryName).EnumerateFileSystemInfos("*.*", SearchOption.AllDirectories)
@@ -39,70 +116,6 @@ public static class IoExtensions
             }
     }
 
-    public static string PathShorten(this string path, int length)
-    {
-        var pathParts = path.Split('\\');
-        var pathBuild = new StringBuilder(path.Length);
-        var lastPart = pathParts[^1];
-        var prevPath = "";
-
-        //Erst prüfen ob der komplette String evtl. bereits kürzer als die Maximallänge ist
-        if (path.Length >= length) return path;
-
-        for (var i = 0; i < pathParts.Length - 1; i++)
-        {
-            pathBuild.Append(pathParts[i] + @"\");
-
-            if ((pathBuild + @"...\" + lastPart).Length >= length) return prevPath;
-
-            prevPath = pathBuild + @"...\" + lastPart;
-        }
-
-        return prevPath;
-    }
-
-
-    public static void Clear(this DirectoryInfo dic)
-    {
-        if (dic == null) throw new ArgumentNullException(nameof(dic));
-
-        if (!dic.Exists) return;
-
-        dic.Attributes = FileAttributes.Normal;
-
-        foreach (var entry in dic.GetFileSystemInfos())
-            switch (entry)
-            {
-                case FileInfo file:
-                    file.Attributes = FileAttributes.Normal;
-                    file.Delete();
-
-                    break;
-                case DirectoryInfo dici:
-                    dic.Attributes = FileAttributes.Normal;
-                    Clear(dici);
-                    dici.Delete();
-
-                    break;
-            }
-    }
-
-    public static void ClearDirectory(this string dic)
-    {
-        if (dic == null) throw new ArgumentNullException(nameof(dic));
-
-        Clear(new DirectoryInfo(dic));
-    }
-
-
-    public static void ClearParentDirectory(this string dic)
-    {
-        if (dic == null) throw new ArgumentNullException(nameof(dic));
-
-        var dic2 = Path.GetDirectoryName(dic);
-        if (!string.IsNullOrWhiteSpace(dic2))
-            ClearDirectory(dic2);
-    }
 
 
     public static string CombinePath(this string path, params string[] paths)
@@ -234,33 +247,9 @@ public static class IoExtensions
     public static IEnumerable<string> EnumerateFiles(this string dic, string filter)
         => Directory.EnumerateFiles(dic, filter, SearchOption.TopDirectoryOnly);
 
-    public static IEnumerable<string> EnumerateTextLinesIfExis(this string path)
-    {
-        if (!File.Exists(path)) yield break;
 
-        using var reader = File.OpenText(path);
 
-        while (true)
-        {
-            var line = reader.ReadLine();
 
-            if (line == null) yield break;
-
-            yield return line;
-        }
-    }
-
-    public static IEnumerable<string> EnumerateTextLines(this TextReader reader)
-    {
-        while (true)
-        {
-            var line = reader.ReadLine();
-
-            if (line == null) yield break;
-
-            yield return line;
-        }
-    }
 
     public static bool ExisDirectory(this string path) => Directory.Exists(path);
 
@@ -432,5 +421,5 @@ public static class IoExtensions
         path.CreateDirectoryIfNotExis();
         #pragma warning restore GU0011
         return new FileStream(path, FileMode.OpenOrCreate, FileAccess.Write, share);
-    }
+    }*/
 }
