@@ -1,4 +1,7 @@
-﻿using JetBrains.Annotations;
+﻿using System;
+using System.IO;
+using JetBrains.Annotations;
+using Tauron.Application.VirtualFiles;
 
 namespace Tauron.Localization;
 
@@ -13,6 +16,13 @@ public enum JsonFileNameMode
 }
 
 [PublicAPI]
-public sealed record JsonConfiguration(
-    string RootDic = "lang", JsonFileNameMode NameMode = JsonFileNameMode.Name,
-    string Fallback = "en");
+public sealed record JsonConfiguration(IDirectory RootDic, JsonFileNameMode NameMode = JsonFileNameMode.Name, string Fallback = "en")
+{
+    public static JsonConfiguration CreateFromApplicationPath(VirtualFileFactory factory,
+        string langDirectory = "lang", JsonFileNameMode nameMode = JsonFileNameMode.Name, string fallback = "en")
+    {
+        var rootPath = Path.Combine(Path.GetDirectoryName(Environment.ProcessPath) ?? Environment.CurrentDirectory, langDirectory);
+
+        return new JsonConfiguration(factory.Local(rootPath), nameMode, fallback);
+    }
+}
