@@ -1,12 +1,13 @@
 ﻿using System;
 using Akka.Util;
 using JetBrains.Annotations;
+using Tauron.Application.VirtualFiles;
 
 namespace Tauron.Temp;
 
 [PublicAPI]
 public interface ITempInfo : IDisposable
 {
-    string FullPath { get; }
+    PathInfo FullPath { get; }
     Option<ITempDic> Parent { get; }
 }
