@@ -1,12 +1,10 @@
-﻿using System.IO;
+﻿using System;
 
 namespace TestApp
 {
     static class Program
     {
         static void Main()
-        {
-            string test = Path.GetFullPath("C:/Test/Test2");
-        }
+        { }
     }
 }
